Merge "build" to "backend"

+++ application/modules/users/controllers/Users.php
@@ -10,7 +10,7 @@ class Users extends MX_Controller {
 		$this->load->model('users/user_model');
 	}
 
-  public function index(){
+  public function index() {
     $data = [
       'user_info' => $this->session->userdata('user_info'),
       'user_types' => $this->get_usertypes()['data'] ?? []
@@ -50,7 +50,7 @@ class Users extends MX_Controller {
     );
   }
 
-  public function load_users(){
+  public function load_users() {
     $searchkey = $this->input->post('searchkey') ?? NULL;
 		$limit = $this->input->post('limit') ?? NULL;
 		$start = $this->input->post('start') ?? NULL;
@@ -90,7 +90,7 @@ class Users extends MX_Controller {
     echo json_encode( $data );
   }
 
-  public function add_new_user(){
+  public function add_new_user() {
     $data['response'] = FALSE;
     $params = format_parameters(clean_parameters(json_decode($this->input->post('params'), true), []));
     if (isset($params['confirmpasswd'])) {
@@ -101,7 +101,7 @@ class Users extends MX_Controller {
 			$result = $this->user_model->add_new_user($params);
       $data['message'] = $result['message'];
 
-			if (!empty($result) && $result['code'] == 0){
+			if (!empty($result) && $result['code'] == 0) {
 				$data['response'] = TRUE;
 				$data['message'] = 'Successfully added new user.';
         $res = $this->update_user_photo(['user_id' => $result['data']['user_id']], FALSE);
@@ -114,7 +114,7 @@ class Users extends MX_Controller {
 		echo json_encode( $data );
   }
 
-  public function update_user($params = [], $ajax = TRUE){
+  public function update_user($params = [], $ajax = TRUE) {
     $data['response'] = FALSE;
     $params = ($ajax) ? $this->input->post('params') : $params;
     $params = format_parameters(clean_parameters($params, []));
@@ -136,7 +136,7 @@ class Users extends MX_Controller {
 			$result = $this->user_model->update_user($id, $params);
       $data['message'] = $result['message'];
 
-			if (!empty($result) && $result['code'] == 0){
+			if (!empty($result) && $result['code'] == 0) {
 				$data['response'] = TRUE;
 				$data['message'] = 'Successfully updated user.';
 			}
@@ -157,6 +157,7 @@ class Users extends MX_Controller {
 
     try {
       $photo = $_FILES['file'] ?? [];
+      $old_photo = $this->input->post('old_photo') ?? '';
       $user_id = $this->input->post('user_id') ?? $params['user_id'] ?? 0;
       $user_id = urldecode($user_id);
 
@@ -172,20 +173,18 @@ class Users extends MX_Controller {
       $allowedExts = ['jpg','jpeg','png','gif','PNG','JPG','JPEG','GIF'];
       $allowedMimes = ['image/jpeg','image/jpg','image/png','image/gif'];
 
-      if (!in_array($ext, $allowedExts) || !in_array($mime, $allowedMimes) || $size > 5) {
-        throw new Exception('Invalid file type or size. Please use image files only with no more than 5MB.');
+      if (!in_array($ext, $allowedExts) || !in_array($mime, $allowedMimes) || $size > MAX_FILESIZE_MB) {
+        throw new Exception('UPDATE USER PHOTO: Invalid file type or size. Please use image files only with no more than 5MB.');
       }
 
-      $newName = decrypt($user_id) . '.' . $ext;
+      $newName = md5(decrypt($user_id) . date('Y-m-d H:i:s A')) . '.' . $ext;
       $source = $photo['tmp_name'];
       $folder = ENV['image_upload_path'] . 'users/';
       $target = $folder . $newName;
 
-      foreach ($allowedExts as $ex) {
-        $filepath = $folder . decrypt($user_id) . '.' . $ex;
-        if (file_exists($filepath)) {
-          unlink($filepath); // delete existing file
-        }
+      $filepath = $folder . $old_photo;
+      if (file_exists($filepath) && !empty($old_photo) && $old_photo != 'default-user.jpg') {
+        unlink($filepath); // delete existing file
       }
 
       if(move_uploaded_file($source, $target)) {
@@ -201,6 +200,7 @@ class Users extends MX_Controller {
         ], FALSE);
 
         $data = $result;
+        $data['data'] = ['user_photo' => $newName];
       }
 
     } catch (Exception $e) {
@@ -214,7 +214,7 @@ class Users extends MX_Controller {
     return $data;
   }
 
-  private function get_usertypes(){
+  private function get_usertypes() {
     $data['response'] = FALSE;
     $data['message'] = 'Failed';
      try {
